pyproject.toml
[tool.poetry]
name = "turtle-shell"
version = "0.0.1"
description = "Easily wrap up your shell and utlity scripts into nice Django UIs/graphQL APIs"
authors = ["Jeff Tratner <jeffrey.tratner@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.8,<4"
pydantic = "*"
graphene-pydantic = "==0.0.9"
graphene-django = ">=2.15.0"
django-crispy-forms = ">=1.11.2"
defopt = ">=6.1.0"
django-filter=">=2.4.0"

[tool.poetry.dev-dependencies]
Werkzeug = "*"
ipdb = "0.13.7"
psycopg2-binary = "2.8.6"
watchgod = "0.6"
mypy = "0.812"
django-stubs = "1.7.0"
pytest = "6.2.3"
pytest-sugar = "0.9.4"
Sphinx = "3.5.3"
sphinx-autobuild = "2021.3.14"
flake8 = "3.9.0"
flake8-isort = "4.0.0"
coverage = "5.5"
black = "20.8b1"
pylint-django = "2.4.3"
pylint-celery = "0.3"
pre-commit = "2.12.0"
factory-boy = "3.2.0"
django-debug-toolbar = "3.2"
django-extensions = "3.1.2"
django-coverage-plugin = "1.8.0"
pytest-django = "4.2.0"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
